+++ app/dashboard/route.ts
@@ -3,7 +3,8 @@ import { inject as service } from '@ember/service';
 
 export default class Dashboard extends Route.extend({
     session: service('session'),
-
+    analytics: service(),
+    currentUser: service('current-user'),
     async beforeModel(transition) {
         await this._super(transition);
 
@@ -12,4 +13,10 @@ export default class Dashboard extends Route.extend({
             this.transitionTo('home');
         }
     },
+    actions: {
+        didTransition(this: Dashboard) {
+            this.get('analytics').trackPage();
+        },
+    },
+
 }) { }