+++ app/guid-user/quickfiles/controller.ts
@@ -1,9 +1,8 @@
 import Controller from '@ember/controller';
 import { computed } from '@ember/object';
 import { inject as service } from '@ember/service';
-import Analytics from 'ember-osf-web/mixins/analytics';
 
-export default class UserQuickfiles extends Controller.extend(Analytics, {
+export default class UserQuickfiles extends Controller.extend({
     actions: {
         async openFile(this: UserQuickfiles, file, show) {
             const guid = file.get('guid') || await file.getGuid();
@@ -11,7 +10,7 @@ export default class UserQuickfiles extends Controller.extend(Analytics, {
         },
     },
 }) {
-    currentUser = service('currentUser');
+    currentUser = service('current-user');
     pageName = 'QuickFiles';
 
     user = computed.alias('model.taskInstance.value');